Refactored Utils
